backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
loguru>=0.7.0

pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
email-validator>=2.0.0
bcrypt==3.2.2
